refactor(storage): move resumable upload functions
We will need these to dynamically switch between resumable and one-shot
uploads.
